Move general data splitting tasks to abstract base class
