+++ great_expectations/execution_engine/sqlalchemy_data_splitter.py
@@ -15,12 +15,15 @@ not by itself.
 """
 
 import datetime
-import enum
-from typing import Callable, List, Union
+from typing import List, Union
 
 from dateutil.parser import parse
 
 from great_expectations.exceptions import exceptions as ge_exceptions
+from great_expectations.execution_engine.split_and_sample.data_splitter import (
+    DataSplitter,
+    DatePart,
+)
 
 try:
     import sqlalchemy as sa
@@ -39,56 +42,12 @@ except ImportError:
     Label = None
 
 
-class DatePart(enum.Enum):
-    """SQL supported date parts for most dialects."""
-
-    YEAR = "year"
-    MONTH = "month"
-    WEEK = "week"
-    DAY = "day"
-    HOUR = "hour"
-    MINUTE = "minute"
-    SECOND = "second"
-
-    def __eq__(self, other):
-        return self.value == other.value
-
-
-class SqlAlchemyDataSplitter:
+class SqlAlchemyDataSplitter(DataSplitter):
     """Methods for splitting data accessible via SqlAlchemyExecutionEngine.
 
     Note, for convenience, you can also access DatePart via the instance variable
     date_part e.g. SqlAlchemyDataSplitter.date_part.MONTH
     """
-
-    date_part: DatePart = DatePart
-
-    def get_splitter_method(self, splitter_method_name: str) -> Callable:
-        """Get the appropriate splitter method from the method name.
-
-        Args:
-            splitter_method_name: name of the splitter to retrieve.
-
-        Returns:
-            splitter method.
-        """
-        splitter_method_name: str = self._get_splitter_method_name(splitter_method_name)
-
-        return getattr(self, splitter_method_name)
-
-    def _get_splitter_method_name(self, splitter_method_name: str) -> str:
-        """Accept splitter methods with or without starting with `_`.
-
-        Args:
-            splitter_method_name: splitter name starting with or without preceding `_`.
-
-        Returns:
-            splitter method name stripped of preceding underscore.
-        """
-        if splitter_method_name.startswith("_"):
-            return splitter_method_name[1:]
-        else:
-            return splitter_method_name
 
     def split_on_year(
         self,
@@ -361,22 +320,6 @@ class SqlAlchemyDataSplitter:
             column_name=column_name,
             date_parts=[DatePart.YEAR, DatePart.MONTH, DatePart.DAY],
         )
-
-    def _convert_date_parts(
-        self, date_parts: Union[List[DatePart], List[str]]
-    ) -> List[DatePart]:
-        """Convert a list of date parts to DatePart objects.
-
-        Args:
-            date_parts: List of DatePart or string representations of DatePart.
-
-        Returns:
-            List of DatePart objects
-        """
-        return [
-            DatePart(date_part.lower()) if isinstance(date_part, str) else date_part
-            for date_part in date_parts
-        ]
 
     def get_split_query_for_data_for_batch_identifiers_for_split_on_date_parts(
         self,
